[FEATURE] Add bubble sort implementation and tests
A new file has been created to house the bubble sort algorithm. This algorithm takes in an array and sorts it. Additional tests were also added to the test_sort file to verify the correctness of this new algorithm. The test_sort file now also uses a pytest.fixture to setup arrays for testing.