[UPDATE] Add use examples, improve imports and bump version to 1.0
In this commit, we've added a new script to compare the execution times of two sorting algorithms: Quick Sort and Selection Sort. The code has been updated to use the algorithms directly from their modules. We've also bumped the version of the project to 1.0. Moreover, a script to demonstrate the use of Quick Sort algorithm in a real-world scenario (sorting offers) has been added.